# use arith-by-dot-products approach for fp32 accumulation in fp16 matmul